refactor(image): ImageResponse 경로 재설정
