+++ src/test/java/com/sprarta/sproutmarket/domain/image/profileImage/controller/ProfileImageControllerTest.java
@@ -4,7 +4,7 @@ import com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper;
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.epages.restdocs.apispec.Schema;
 import com.sprarta.sproutmarket.domain.CommonMockMvcControllerTestSetUp;
-import com.sprarta.sproutmarket.domain.image.dto.ImageResponse;
+import com.sprarta.sproutmarket.domain.image.dto.response.ImageResponse;
 import com.sprarta.sproutmarket.domain.image.profileImage.service.ProfileImageService;
 import com.sprarta.sproutmarket.domain.item.dto.request.ImageNameRequest;
 import com.sprarta.sproutmarket.domain.user.entity.CustomUserDetails;
